ZOOKEEPER-2641

+++ zookeeper-server/src/main/java/org/apache/zookeeper/server/ServerStats.java
@@ -20,13 +20,12 @@ package org.apache.zookeeper.server;
 
 
 
+import java.util.concurrent.atomic.AtomicLong;
 import org.apache.zookeeper.common.Time;
 import org.apache.zookeeper.server.metric.AvgMinMaxCounter;
 import org.apache.zookeeper.server.quorum.BufferStats;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import java.util.concurrent.atomic.AtomicLong;
 
 /**
  * Basic Server Statistics
@@ -64,7 +63,7 @@ public class ServerStats {
         return requestLatency.getMin();
     }
 
-    public long getAvgLatency() {
+    public double getAvgLatency() {
         return requestLatency.getAvg();
     }
